requirements.txt
# Core
fastapi
uvicorn[standard]
gunicorn

# Templates & Static
jinja2
python-multipart

# Image Processing
opencv-python-headless
mediapipe
numpy

# Payment
stripe

# Database
sqlalchemy[asyncio]
asyncpg

# Auth & Utils
PyJWT
python-dotenv

# HTTP Client (for PhotoRoom API)
httpx
